[Feat]: CoreDataManager 구현 [Refactor]: AppDelegate, SceneDelegate - CoreDataManager의 싱글턴 객체를 이용하여 수정
