+++ Todoey/Todoey/SceneDelegate.swift
@@ -79,7 +79,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         print("SceneDelegate: UI Lifecycle - sceneDidEnterBackground")
         
         // Save changes in the application's managed object context when the application transitions to the background.
-        (UIApplication.shared.delegate as? AppDelegate)?.saveContext()
+        (UIApplication.shared.delegate as? AppDelegate)?.coreDataManager.saveContext()
     }
